\name{catchstatus-package}
\alias{catchstatus-package}
\docType{package}
\title{Methods to assign status to fish stocks based on catch series.}
\description{Functions to assign status to fish stocks based on catch series. Currently implements Froese and Kesner-Reyes 2002 (ICES Council Meeting Report) and Froese et al. 2012 (Mar. Biol.) updated algorithm. Includes functions to plot the output. Will include functions to test the accuracy and bias of the classifications.}
\details{
\tabular{ll}{Package: \tab catchstatus\cr
Type: \tab Package\cr
Title: \tab Methods to assign status to fish stocks based on catch series.\cr
Version: \tab 0.2\cr
Date: \tab 2012-03-29\cr
Author: \tab Sean Anderson\cr
Maintainer: \tab Sean Anderson <sean@seananderson.ca>\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr}
}
\author{Sean Anderson}

\keyword{ package }


